Update tests related to listing repositories
