--- tests/integration/Repositories/GetRepositoryListTest.cs
@@ -1,57 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Threading.Tasks;
-
-namespace Laserfiche.Repository.Api.Client.IntegrationTest.Repositories
-{
-    [TestClass]
-    public class GetRepositoryListTest : BaseTest
-    {
-        [TestInitialize]
-        public void Initialize()
-        {
-            client = CreateClient();
-        }
-
-        [TestMethod]
-        public async Task GetRepositoryList_ReturnSuccessful()
-        {
-            var result = await client.RepositoriesClient.GetRepositoryListAsync().ConfigureAwait(false);
-            Assert.IsTrue(result.Count > 0, "No repositories found.");
-
-            bool foundRepo = false;
-            foreach (var repoInfo in result)
-            {
-                Assert.IsFalse(string.IsNullOrEmpty(repoInfo.RepoId));
-                if (AuthorizationType != AuthorizationType.API_SERVER_USERNAME_PASSWORD)
-                {
-                    Assert.IsFalse(string.IsNullOrEmpty(repoInfo.WebclientUrl));
-                    Assert.IsTrue(repoInfo.WebclientUrl.Contains(repoInfo.RepoId));
-                }
-
-                if (repoInfo.RepoId.Equals(RepositoryId, System.StringComparison.OrdinalIgnoreCase))
-                    foundRepo = true;
-            }
-            Assert.IsTrue(foundRepo);
-        }
-
-        [TestMethod]
-        public async Task GetSelfHostedRepositoryList_ReturnSuccessful()
-        {
-            if (AuthorizationType == AuthorizationType.CLOUD_ACCESS_KEY)
-            {
-                return; // There's no point testing if it is a cloud environment
-            }
-            var result = await RepositoriesClient.GetSelfHostedRepositoryListAsync(BaseUrl).ConfigureAwait(false);
-            Assert.IsTrue(result.Count > 0, "No repositories found.");
-            Assert.IsNotNull(result);
-            bool foundRepo = false;
-            foreach (var repoInfo in result)
-            {
-                Assert.IsFalse(string.IsNullOrEmpty(repoInfo.RepoId));
-                if (repoInfo.RepoId.Equals(RepositoryId, System.StringComparison.OrdinalIgnoreCase))
-                    foundRepo = true;
-            }
-            Assert.IsTrue(foundRepo);
-        }
-    }
-}
